+++ apps/auth/src/app/auth/access-token-payload.interface.ts
@@ -1,0 +1,8 @@
+import { UserRole } from '@soer/sr-common-interfaces';
+
+export interface AccessTokenPayload {
+  id: number;
+  uuid: string;
+  email: string;
+  userRole: UserRole;
+}