+++ apps/auth/src/app/auth/refresh-token-payload.interface.ts
@@ -1,0 +1,7 @@
+import { UserRole } from '@soer/sr-common-interfaces';
+
+export interface RefreshTokenPayload {
+  userId: number;
+  userEmail: string;
+  userRole: UserRole;
+}