+++ apps/naris-api/src/app/modules/json/json.entity.ts
@@ -8,11 +8,11 @@ export class JsonEntity {
   id: number;
 
   @ApiProperty()
-  @CreateDateColumn({ type: 'timestamp', default: () => 'CURRENT_TIMESTAMP(6)' })
+  @CreateDateColumn({ type: 'datetime', default: () => 'CURRENT_TIMESTAMP(6)' })
   createdAt: Date;
 
   @ApiProperty()
-  @UpdateDateColumn({ type: 'timestamp', default: () => 'CURRENT_TIMESTAMP(6)', onUpdate: 'CURRENT_TIMESTAMP(6)' })
+  @UpdateDateColumn({ type: 'datetime', default: () => 'CURRENT_TIMESTAMP(6)', onUpdate: 'CURRENT_TIMESTAMP(6)' })
   updatedAt: Date;
 
   @ApiProperty()
@@ -28,6 +28,6 @@ export class JsonEntity {
   author_email: string;
 
   @ApiProperty()
-  @Column({ default: () => 'PRIVATE' })
+  @Column({ default: 'PRIVATE' })
   accessTag: string;
 }