+++ apps/naris/.storybook/__mocks__/environment.ts
@@ -1,8 +1,11 @@
 import { dev } from '../../src/environments/feature-flags';
 import { EnvironmentInterface } from '../../src/environments/environment.interface';
+import { getDevHostVariables } from '../../src/environments/getHostVariables';
 
 const HOST = '';
+
 export const environment: EnvironmentInterface = {
+  ...getDevHostVariables(),
   production: false,
   host: HOST,
   googleAuthUrl: HOST + '/api/auth/google',
@@ -12,5 +15,7 @@ export const environment: EnvironmentInterface = {
   assetsUrl: HOST + '/assets/',
   privateAssetsUrl: HOST + '/assets/private/',
   payServiceUrl: HOST + '/api/v2/seller',
-  features: dev,
+  features: {
+    auth_v2: true,
+  },
 };